# docs/api-documents.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LinkUp API",
    "version": "1.0.0",
    "description": "API documentation for LinkUp",
    "contact": {
      "name": "LinkUp",
      "url": "api-linkup.id.vn"
    }
  },
  "servers": [
    {
      "url": "api-linkup.id.vn",
      "description": "Vercel Server"
    }
  ],
  "paths": {
...
    "/api/media/getPost": {
      "get": {
        "summary": "Lấy danh sách bài viết",
        "tags": ["Media"],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "ID của người dùng hiện tại"
          }
        ],
        "responses": {
          "200": { "description": "Lấy danh sách bài viết thành công" },
          "400": { "description": "Lỗi không xác định" }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/media/createPost": {
      "post": {
        "summary": "Tạo bài viết mới",
        "tags": ["Media"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "image": { "type": "string" },
                  "userId": { "type": "int" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tạo bài viết thành công" },
          "400": { "description": "Lỗi đầu vào không hợp lệ" }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/like/getLikes": {
      "get": {
        "summary": "Lấy danh sách lượt thích",
        "tags": ["Like"],
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "schema": {
              "type": "int"
            },
            "description": "ID của bài viết"
          }
        ],
        "responses": {
          "200": { "description": "Lấy danh sách comment thành công" },
          "400": { "description": "Lỗi không xác định" }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/like/createLike": {
      "post": {
        "summary": "Tạo bình luận mới cho bài viết",
        "tags": ["Comment"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "int" },
                  "postId": { "type": "int" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tạo lượt thích thành công" },
          "400": { "description": "Lỗi đầu vào không hợp lệ" }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Người dùng cần truyền Access Token vào phần Authorization theo định dạng Bearer Token. Access Token này được trả về khi đăng nhập thành công."
      }
    }
  }
}